+++ packages/frontend/src/store/analyzer/saga/runAnalyzer/runAnalyzer.saga.spec.ts
@@ -1,15 +1,7 @@
 import { expectSaga } from 'redux-saga-test-plan'
 import * as matchers from 'redux-saga-test-plan/matchers'
 import { combineReducers } from 'redux'
-import type {
-  TAbis,
-  TByteCodeMap,
-  TMappedContractNames,
-  TMappedSourceCodes,
-  TMappedSourceMap,
-  TStepInstrctionsMap,
-  TTransactionData,
-} from '@evm-debuger/types'
+import type { TAbis, TByteCodeMap, TMappedContractNames, TMappedSourceMap, TStepInstrctionsMap, TTransactionData } from '@evm-debuger/types'
 
 import { analyzerActions, analyzerReducer } from '../../analyzer.slice'
 import { AnalyzerState, analyzerStagesAdapter } from '../../analyzer.state'